build 34 - Fixed non-integer crosshair scales not positioning correctly.

+++ trunk/paintball2/client/screen.h
@@ -48,7 +48,7 @@ extern	cvar_t		*crosshair;
 extern	vrect_t		scr_vrect;		// position of render window
 
 extern	char		crosshair_pic[MAX_QPATH];
-extern	int			crosshair_width, crosshair_height;
+extern	float		crosshair_width, crosshair_height;
 
 void SCR_AddDirtyPoint (int x, int y);
 void SCR_DirtyScreen (void);
